ovltest: Don't clean writeback connector's crtc when disable it
If we set a writeback connector's crtc to NULl when we want to
disable it, the drm core will mark the crtc's connectors_changed
as true, this will triger a disable/enable for crtc-->other
connector(hdmi/dp/edp/). This is not we want to see.
So we only set a writeback connector's FB to NULL, when we want
disable a writeback connector.

Signed-off-by: Andy Yan <andyshrk@gmail.com>